requirements.txt
aiofiles==24.1.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
black==24.10.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.115.0
fastapi-sessions==0.3.2
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pillow==11.0.0
platformdirs==4.8.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==5.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.3
redis==7.1.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
